# artifacts/api-server/src/routes/domains.ts
import { Router } from "express";
import type { Request, Response } from "express";
import { parseIdParam, parsePagination } from "../lib/params";
import { db } from "@workspace/db";
import { domainsTable, usersTable, auditLogsTable } from "@workspace/db";
import { CreateDomainBody, UpdateDomainBody } from "@workspace/api-zod";
import { eq, desc, isNull, isNotNull, and, gte, count } from "drizzle-orm";
import { logAudit } from "../lib/audit";
import { sendError } from "../lib/errors";

async function resolveOwnerName(ownerId: number | null | undefined): Promise<string | null> {
  if (!ownerId) return null;
  const [user] = await db.select({ name: usersTable.name }).from(usersTable).where(eq(usersTable.id, ownerId));
  return user?.name ?? null;
}

function fmt(d: typeof domainsTable.$inferSelect, ownerName: string | null = null) {
  return {
    ...d,
    ownerName,
    createdAt: d.createdAt.toISOString(),
    updatedAt: d.updatedAt.toISOString(),
    deletedAt: d.deletedAt ? d.deletedAt.toISOString() : null,
  };
}

const router = Router();

router.get("/expiring", async (req: Request, res: Response) => {
  try {
    const all = await db
      .select({ domain: domainsTable, ownerName: usersTable.name })
      .from(domainsTable)
      .leftJoin(usersTable, eq(domainsTable.ownerId, usersTable.id))
      .where(isNull(domainsTable.deletedAt));
    const now = new Date();
    const in90Days = new Date(now.getTime() + 90 * 24 * 60 * 60 * 1000);
    const expiring = all.filter(({ domain: d }) => {
      const expiry = d.sslExpiry || d.registrationExpiry;
      if (!expiry) return false;
      const date = new Date(expiry);
      return date >= now && date <= in90Days;
    });
    return res.json(expiring.map(({ domain: d, ownerName }) => fmt(d, ownerName ?? null)));
  } catch (err) {
    req.log.error({ err }, "Error fetching expiring domains");
    return sendError(res, 500, "Internal server error");
  }
});

router.get("/", async (req: Request, res: Response) => {
  try {
    const { limit, offset } = parsePagination(req);
    const [[{ total }], results] = await Promise.all([
      db.select({ total: count() }).from(domainsTable).where(isNull(domainsTable.deletedAt)),
      db.select({ domain: domainsTable, ownerName: usersTable.name })
        .from(domainsTable)
        .leftJoin(usersTable, eq(domainsTable.ownerId, usersTable.id))
        .where(isNull(domainsTable.deletedAt))
        .limit(limit)
        .offset(offset),
    ]);
    return res.json({ data: results.map(({ domain: d, ownerName }) => fmt(d, ownerName ?? null)), total: Number(total) });
  } catch (err) {
    req.log.error({ err }, "Error listing domains");
    return sendError(res, 500, "Internal server error");
  }
});

router.post("/", async (req: Request, res: Response) => {
  try {
    const body = CreateDomainBody.parse(req.body);
    const [item] = await db.insert(domainsTable).values(body).returning();
    const ownerName = await resolveOwnerName(item.ownerId);
    await logAudit(req, "CREATE", "Domain", item.id, item.name);
    return res.status(201).json(fmt(item, ownerName));
  } catch (err) {
    req.log.error({ err }, "Error creating domain");
    return sendError(res, 400, "Invalid request");
  }
});

router.post("/:id/restore", async (req: Request, res: Response) => {
  try {
    const id = parseIdParam(req);
    const cutoff = new Date(Date.now() - 30 * 24 * 60 * 60 * 1000);
    const [item] = await db.update(domainsTable)
      .set({ deletedAt: null, updatedAt: new Date() })
      .where(and(eq(domainsTable.id, id), isNotNull(domainsTable.deletedAt), gte(domainsTable.deletedAt, cutoff)))
      .returning();
    if (!item) return sendError(res, 404, "Not found or outside the 30-day restore window");
    const ownerName = await resolveOwnerName(item.ownerId);
    await logAudit(req, "RESTORE", "Domain", item.id, item.name);
    return res.json(fmt(item, ownerName));
  } catch (err) {
    req.log.error({ err }, "Error restoring domain");
    return sendError(res, 500, "Internal server error");
  }
});

router.get("/:id/history", async (req: Request, res: Response) => {
  try {
    const id = parseIdParam(req);
    const entries = await db
      .select()
      .from(auditLogsTable)
      .where(eq(auditLogsTable.entityId, id))
      .orderBy(desc(auditLogsTable.createdAt));
    const domainEntries = entries.filter(e => e.entityType === "Domain");
    return res.json(
      domainEntries.map(e => ({
        id: e.id,
        action: e.action,
        entityName: e.entityName,
        userName: e.userName,
        changes: e.changes,
        createdAt: e.createdAt.toISOString(),
      }))
    );
  } catch (err) {
    req.log.error({ err }, "Error fetching domain history");
    return sendError(res, 500, "Internal server error");
  }
});

router.get("/:id", async (req: Request, res: Response) => {
  try {
    const id = parseIdParam(req);
    const [result] = await db
      .select({ domain: domainsTable, ownerName: usersTable.name })
      .from(domainsTable)
      .leftJoin(usersTable, eq(domainsTable.ownerId, usersTable.id))
      .where(and(eq(domainsTable.id, id), isNull(domainsTable.deletedAt)));
    if (!result) return sendError(res, 404, "Not found");
    return res.json(fmt(result.domain, result.ownerName ?? null));
  } catch (err) {
    req.log.error({ err }, "Error fetching domain");
    return sendError(res, 500, "Internal server error");
  }
});

router.patch("/:id", async (req: Request, res: Response) => {
  try {
    const id = parseIdParam(req);
    const body = UpdateDomainBody.parse(req.body);
    const [item] = await db.update(domainsTable)
      .set({ ...body, updatedAt: new Date() })
      .where(and(eq(domainsTable.id, id), isNull(domainsTable.deletedAt)))
      .returning();
    if (!item) return sendError(res, 404, "Not found");
    const ownerName = await resolveOwnerName(item.ownerId);
    return res.json(fmt(item, ownerName));
  } catch (err) {
    req.log.error({ err }, "Error updating domain");
    return sendError(res, 400, "Invalid request");
  }
});

router.delete("/:id", async (req: Request, res: Response) => {
  try {
    const id = parseIdParam(req);
    const [item] = await db.update(domainsTable)
      .set({ deletedAt: new Date(), updatedAt: new Date() })
      .where(and(eq(domainsTable.id, id), isNull(domainsTable.deletedAt)))
      .returning();
    if (!item) return sendError(res, 404, "Not found");
    await logAudit(req, "DELETE", "Domain", item.id, item.name);
    return res.status(204).send();
  } catch (err) {
    req.log.error({ err }, "Error deleting domain");
    return sendError(res, 500, "Internal server error");
  }
});

export default router;