# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: MK Digital Operations Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /infrastructure:
    get:
      operationId: listInfrastructure
      tags: [infrastructure]
      summary: List infrastructure items
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Infrastructure list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedInfrastructureList"
    post:
      operationId: createInfrastructure
      tags: [infrastructure]
      summary: Create infrastructure item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InfrastructureInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfrastructureItem"
  /infrastructure/{id}/dependents:
    get:
      operationId: getInfrastructureDependents
      tags: [infrastructure]
      summary: Get counts of applications dependent on this infrastructure item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dependent record counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfrastructureDependents"
        "404":
          description: Not found
  /infrastructure/{id}:
    get:
      operationId: getInfrastructure
      tags: [infrastructure]
      summary: Get infrastructure item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Infrastructure item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfrastructureItem"
    patch:
      operationId: updateInfrastructure
      tags: [infrastructure]
      summary: Update infrastructure item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InfrastructureUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfrastructureItem"
    delete:
      operationId: deleteInfrastructure
      tags: [infrastructure]
      summary: Delete infrastructure item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /infrastructure/{id}/restore:
    post:
      operationId: restoreInfrastructure
      tags: [infrastructure]
      summary: Restore a soft-deleted infrastructure item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restored infrastructure item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfrastructureItem"
        "404":
          description: Not found

  # ─── Databases ────────────────────────────────────────────────────────────
...
  /domains:
    get:
      operationId: listDomains
      tags: [domains]
      summary: List domains
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Domains
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedDomainList"
    post:
      operationId: createDomain
      tags: [domains]
      summary: Create domain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DomainInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
  /domains/{id}:
    get:
      operationId: getDomain
      tags: [domains]
      summary: Get domain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Domain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
    patch:
      operationId: updateDomain
      tags: [domains]
      summary: Update domain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DomainUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
    delete:
      operationId: deleteDomain
      tags: [domains]
      summary: Delete domain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /domains/{id}/restore:
    post:
      operationId: restoreDomain
      tags: [domains]
      summary: Restore a soft-deleted domain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restored domain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
        "404":
          description: Not found
  /domains/{id}/history:
    get:
      operationId: getDomainHistory
      tags: [domains]
      summary: Get audit/change history for a domain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: History entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DomainHistoryEntry"
...
components:
  schemas:
...
    InfrastructureDependents:
      type: object
      properties:
        applications:
          type: integer
        total:
          type: integer
      required: [applications, total]
...
    InfrastructureItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        provider:
          type: ["string", "null"]
        status:
          type: string
        ipAddress:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        cpuCores:
          type: ["integer", "null"]
        ramGb:
          type: ["integer", "null"]
        diskGb:
          type: ["integer", "null"]
        os:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        teamId:
          type: ["integer", "null"]
        patchStatus:
          type: string
        lastPatchedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        deletedAt:
          type: ["string", "null"]
      required: [id, name, type, status, patchStatus, createdAt, updatedAt]
    InfrastructureInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
        provider:
          type: string
        status:
          type: string
        ipAddress:
          type: string
        location:
          type: string
        cpuCores:
          type: integer
        ramGb:
          type: integer
        diskGb:
          type: integer
        os:
          type: string
        notes:
          type: string
        ownerId:
          type: ["integer", "null"]
        teamId:
          type: ["integer", "null"]
        patchStatus:
          type: string
        lastPatchedAt:
          type: string
      required: [name, type, status]
    InfrastructureUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        provider:
          type: string
        status:
          type: string
        ipAddress:
          type: string
        location:
          type: string
        cpuCores:
          type: integer
        ramGb:
          type: integer
        diskGb:
          type: integer
        os:
          type: string
        notes:
          type: string
        ownerId:
          type: ["integer", "null"]
        teamId:
          type: ["integer", "null"]
        patchStatus:
          type: string
        lastPatchedAt:
          type: string

    # ── Databases ─────────────────────────────────────────────────────────
...
    Domain:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        registrar:
          type: ["string", "null"]
        registrationExpiry:
          type: ["string", "null"]
        sslProvider:
          type: ["string", "null"]
        sslExpiry:
          type: ["string", "null"]
        sslStatus:
          type: string
        dnsProvider:
          type: ["string", "null"]
        cloudflarEnabled:
          type: boolean
        status:
          type: string
        applicationId:
          type: ["integer", "null"]
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        teamId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        deletedAt:
          type: ["string", "null"]
      required: [id, name, sslStatus, cloudflarEnabled, status, createdAt, updatedAt]
    DomainInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        registrar:
          type: string
        registrationExpiry:
          type: string
        sslProvider:
          type: string
        sslExpiry:
          type: string
        sslStatus:
          type: string
        dnsProvider:
          type: string
        cloudflarEnabled:
          type: boolean
        status:
          type: string
        applicationId:
          type: integer
        ownerId:
          type: ["integer", "null"]
        teamId:
          type: ["integer", "null"]
        notes:
          type: string
      required: [name, status]
    DomainUpdate:
      type: object
      properties:
        name:
          type: string
        registrar:
          type: string
        registrationExpiry:
          type: string
        sslProvider:
          type: string
        sslExpiry:
          type: string
        sslStatus:
          type: string
        dnsProvider:
          type: string
        cloudflarEnabled:
          type: boolean
        status:
          type: string
        applicationId:
          type: integer
        ownerId:
          type: ["integer", "null"]
        teamId:
          type: ["integer", "null"]
        notes:
          type: string
    DomainHistoryEntry:
      type: object
      properties:
        id:
          type: integer
        action:
          type: string
        entityName:
          type: ["string", "null"]
        userName:
          type: ["string", "null"]
        changes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, action, createdAt]

    # ── Repositories ──────────────────────────────────────────────────────
...
    PaginatedInfrastructureList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/InfrastructureItem"
        total:
          type: integer
...
    PaginatedDomainList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Domain"
        total:
          type: integer
...