# lib/api-zod/src/generated/types/domain.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MK Digital Operations Center API
 * OpenAPI spec version: 1.0.0
 */

export interface Domain {
  id: number;
  name: string;
  /** @nullable */
  registrar?: string | null;
  /** @nullable */
  registrationExpiry?: string | null;
  /** @nullable */
  sslProvider?: string | null;
  /** @nullable */
  sslExpiry?: string | null;
  sslStatus: string;
  /** @nullable */
  dnsProvider?: string | null;
  cloudflarEnabled: boolean;
  status: string;
  /** @nullable */
  applicationId?: number | null;
  /** @nullable */
  ownerId?: number | null;
  /** @nullable */
  ownerName?: string | null;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  deletedAt?: string | null;
}