# lib/api-zod/src/generated/types/infrastructureUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MK Digital Operations Center API
 * OpenAPI spec version: 1.0.0
 */

export interface InfrastructureUpdate {
  name?: string;
  type?: string;
  provider?: string;
  status?: string;
  ipAddress?: string;
  location?: string;
  cpuCores?: number;
  ramGb?: number;
  diskGb?: number;
  os?: string;
  notes?: string;
  /** @nullable */
  ownerId?: number | null;
  /** @nullable */
  teamId?: number | null;
  patchStatus?: string;
  lastPatchedAt?: string;
}